+++ dart/constraint/PGSLCPSolver.cpp
@@ -41,7 +41,7 @@
 #include "dart/constraint/ConstrainedGroup.hpp"
 #include "dart/constraint/ConstraintBase.hpp"
 #include "dart/external/odelcpsolver/lcp.h"
-#include "dart/lcpsolver/Lemke.hpp"
+#include "dart/math/Lemke.hpp"
 
 namespace dart {
 namespace constraint {