+++ dart/math/Lemke.cpp
@@ -30,7 +30,7 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "dart/lcpsolver/Lemke.hpp"
+#include "dart/math/Lemke.hpp"
 
 #include "dart/math/Helpers.hpp"
 
@@ -40,7 +40,7 @@
 #include <cmath>
 
 namespace dart {
-namespace lcpsolver {
+namespace math {
 
 // double RandDouble(double _low, double _high) {
 //  double temp;
@@ -356,5 +356,5 @@ bool validate(
   return true;
 }
 
-} // namespace lcpsolver
+} // namespace math
 } // namespace dart